requirements.txt
spotipy
tidalapi
python-dotenv